move gradients around
