+++ MetaMorpheus/GuiFunctions/MetaDraw/BioPolymerCoverage/BioPolymerCoverageMapViewModel.cs
@@ -1,4 +1,5 @@
 using CsvHelper.Configuration.Attributes;
+using GuiFunctions.MetaDraw.BioPolymerCoverage.ColorMapping.Gradient;
 using System;
 using System.Collections.Generic;
 using System.Linq;
