+++ MetaMorpheus/GuiFunctions/MetaDraw/BioPolymerCoverage/ColorMapping/BioPolymerCoverageColorScale.cs
@@ -1,3 +1,4 @@
+using GuiFunctions.MetaDraw.BioPolymerCoverage.ColorMapping.Gradient;
 using System;
 
 namespace GuiFunctions.MetaDraw;
