--- MetaMorpheus/GuiFunctions/MetaDraw/BioPolymerCoverage/ColorMapping/ColorGradientType.cs
@@ -1,10 +1,0 @@
-namespace GuiFunctions.MetaDraw;
-
-public enum ColorGradientType
-{
-    Viridis,
-    Plasma,
-    Inferno,
-    Turbo,
-    Grayscale,
-}
