+++ MetaMorpheus/GuiFunctions/MetaDraw/BioPolymerCoverage/ColorMapping/Gradient/ColorGradientType.cs
@@ -1,0 +1,10 @@
+namespace GuiFunctions.MetaDraw.BioPolymerCoverage.ColorMapping.Gradient;
+
+public enum ColorGradientType
+{
+    Viridis,
+    Plasma,
+    Inferno,
+    Turbo,
+    Grayscale,
+}
