+++ MetaMorpheus/GuiFunctions/MetaDraw/BioPolymerCoverage/ColorMapping/NumericBioPolymerCoverageColorMapper.cs
@@ -1,3 +1,4 @@
+using GuiFunctions.MetaDraw.BioPolymerCoverage.ColorMapping.Gradient;
 using System;
 using System.Collections.Generic;
 using System.Linq;
