+++ MetaMorpheus/Test/MetaDraw/BioPolymerCoverageColorMapperFactoryTests.cs
@@ -1,5 +1,6 @@
 using GuiFunctions;
 using GuiFunctions.MetaDraw;
+using GuiFunctions.MetaDraw.BioPolymerCoverage.ColorMapping.Gradient;
 using NUnit.Framework;
 using System;
 using System.Windows.Media;
