+++ MetaMorpheus/Test/MetaDraw/ColorGradientTests.cs
@@ -1,4 +1,4 @@
-using GuiFunctions.MetaDraw;
+using GuiFunctions.MetaDraw.BioPolymerCoverage.ColorMapping.Gradient;
 using NUnit.Framework;
 using System;
 using System.Linq;
